feat(testing): Add MPI and non-MPI modes to GTest
This commit introduces a new feature that allows tests to be run in either MPI or non-MPI mode. This is controlled by the CMake variable MPI_TEST.

When MPI_TEST is set to ON, the tests are run using multiple processes with MPI. When MPI_TEST is set to OFF, the tests are run in a single process without MPI.

This feature provides flexibility for testing in different environments and can help to identify issues that only occur in multi-process scenarios. It also makes it easier to run tests on systems where MPI is not available.