Enhance ChaseMpi constructor for generality
In this commit, we've added ritzv_ = dla_->get_Ritzv() and resid_ = dla_->get_Resids()  to the ChaseMpi constructor. This enhances its generality and allows it to handle more diverse scenarios. During unit testing, these methods will return NULL which is acceptable as we're not concerned with their return values in the context of these tests.