Add CUDA kernel for computing the sum of absolute values of diagonal entries
This commit introduces a CUDA kernel to efficiently compute the sum of absolute values of diagonal entries in a matrix. This kernel is utilized to compute the shift value required when performing shifted-CholQR decomposition. A virtual function `computeDiagonalAbsSum` is added in `ChaseMpiDLAInterface` class, and its inhertied classes include the actual implementations of this operation.