Lanczos tests added + seg fault issue resolved
The Lanczos internal function now takes a point to a MatrixType type
object for the QuasiHermtian case. This should be generalized to all
type for Matrices. Therefore, the Lanczos function should be templated.

The lanczos for quasi-hermitian matrices seem to work fine for the tiny
matrix of the test. However, the test fails for the larger matrix of
size n = 200. In fact, the ritz values appear in triplet and match every
three eigenvalues of the matrix. This seem to be a minor issue and
should not impact the density approximation. Moreover, this issue
happens in the corner case where M = n and numvec = 1.

TO DO :
- Abstract Lanczos, and pass the pointer to MatrixType even in the
  Hermitian case.
- Add Tests for larger numvec and smaller M for quasi-Hem Lanczos.
- Fix the bug of triplet appearance of ritz values.
- Start the GPU implementation or go for Rayleigh-Ritz?