Multi Type tests for Lanczos on QH mat
The tests for Lanczos on Quasi-Hermitian matrices appear in a separated
file, and work with float, double, complex<float> and complex<double>.
Test matrices are added in the BSE_matrices folder, and accessible for
furter tests.

The issues reported on certain ritz values should probably not impact
the DOS calculation for ChASE. Extreme eigenvalues are accurates.  We
will return to that case later if necesarry.

TO DO :
- Start Rayleigh Ritz on CPU for QH mat
- Fix eventual bugs for the tests on QH Lanczos