Sequential CPU version works!
The sequential CPU version of ChASE works. The algortihm file have been
slightly updated based on hermicity / quasi-hermicity of the matrix.

The lowerb parameter is artificialy decreased in the QH case. The
estimation is probably not accurate enough, causing the polynomial
to filter target eigenvectors from the subspace... Yet, we multiply
lowerb by 0.9 to avoid this issue.

The ritz values do not appear in order at the end of GEEV. A sorting
step seems necessary. We tried sorting the eigenvalues only at the end
but it apparently breaks the convergence... We need to figure out why.

Lastly, since the ritz values are not sorted after GEEV, we loose track
of residLast. This may cause troubles for the optimization of the
polynomial.

TO DO :
- Understand why the sorting algorithm is necessary
- Improve the estimation of lowerb
- Work on the polynomial optimization in the QH case
- Start implementing the GPU version