ChASE GPU : Flip sign cuda kernels + MatrixType
Flip sign kernels have been added. The MatrixType now templates the
ChASEGPU class. The Hmat_ object now points to a MatrixType object, and
should be deleted appropriately.

We may prefer to reconsider the type of the other matrix object for
homogeoneity (yet we combine pointers (Hmat_) with local objects (like Vec1_,
not very consistent...).

TO DO :
- Delete Hmat_ object appropriately
- Check why the CheckPseudoHerm function returns True yet.
- Continue the development for the GPU version of ChASE.