CUDA Kernels for Rayleigh-Ritz added with Unitests
The CUDA Kernels for Rayleigh-Ritz on GPU have been added with their
Unitests. We will probably need to change the filename of shiftDiagonal
which now contains a lot more than shifting the diagonal.

The Rayleigh-Ritz function only misses Xgeev and a proper sorting kernel
on GPU to work.

TO DO :
- Add Xgeev [requires to custom the cmake flags appropriately]
- Add a kernel for sorting ritzvalues and ritzvectors
- Add Unitests for Rayleigh-Ritz GPU
- Update the orthonormalization step for GPU