fixed double free issue in the two tests.
The issue was that:
There is a global mpi_grid for all tests, and in each test itself, it creates a new MPI grid for each test, while we already have a global MPI grid. When these local MPI grids go out of scope at the end of each test, they're being destroyed, but then the global one is also being destroyed at the end of all tests, leading to double free errors.