Tracking of the resid last for QH mat done.
The commit enables the tracking of the residuals in the quasi-hermitian
case. First, we compute the permutation indices for the residuals sorted
in ascending order. Secondly, we compute the permutation indices of the
ritz values sorted in ascending order which are part of the 95% smallest
residuals. These two steps work as a centrifuge since it eliminates the
few ghost points caracterized by larger residuals. Note that we only
permute the indices but not the actual values, since it would have
required moving the ritz vectors and therefore led to more data copies.

However, we sort the residLast EXPLICITELY.

The lowerb bound is always multiplied with the DecayRate, just for
experiments. It may be better not to do it.

TO DO :
- Test on Juwels Booster.
- Fix the instability of HEEVD. In this implementation, we need to
  invert (either explicitly or implicitly) the Q'SQ matrix. If Q is
  almost rank deficient, then M becomes singular and HEEVD breaks. One
  can call the former Rayleigh-Ritz in case of failure to prevent these
  corner cases. The rank deficiency can happen in case of bad parameters
  lambda and lowerb.