Activate GEEV + Improving Cheb. Params + Norma.
In this commit, GEEV can be activated in external programs by defining
the macro XGEEV_EXISTS.

In addition, some bugs related to the sorting of the indices of the
ritz values / residuals have been fixed, which improve in the
calculation for the chebyshev parameters lambda and lowerb.

The ritz vectors are now normalized prior to the back transformation
V = QW.

Lastly, the sorting phase following HEEVD appears useless and has
therefore been deleted. The same logic should apply to GEEV but bugs
appear when removing the sorting here... To be fixed.