Options for saving residuals and double for QR/RR
This commit enables two new options for saving the residuals, and also
for operating QR and RR in double precision instead of single precision.

This commit also fixes a few issues related to the tracking of the
residuals, as well as heuristics for the estimation of lowerb.

In particular, we noticed very fast convergence when setting lowerb = 1,
but at the price of unstability for certain corner cases. Although it
generally leads to one or two more iterations, we recommand enlarging
the filter interval by setting lowerb = 0.98.